### 2025-01-17 66b29a0
Refactor to cleanup map_blocks call
