### 2025-02-20 15cdb52
Add support for Path objects
Closes #35
