### 2025-02-21 b77f101
Set spatial_partitions on Raster.to_points result
Closes #32

This also fixes Raster.get_chunk_bounding_boxes which broke when
chunking of size 1 was used.
