### 2025-03-31 881656f
Add util funcs for comparing null values
