### 2025-04-01 049494d
Improve pathlib support
