+++ tests/test_raster.py
@@ -65,6 +65,7 @@ from tests.utils import (
     arange_nd,
     arange_raster,
     assert_dataarrays_similar,
+    assert_rasters_equal,
     assert_rasters_similar,
     assert_valid_raster,
 )
@@ -662,6 +663,26 @@ def test_raster_from_file(path):
 def test_raster_from_any_errors(data, error_type):
     with pytest.raises(error_type):
         Raster(data)
+
+
+@pytest.fixture(params=[True, False])
+def boolean(request):
+    return request.param
+
+
+@pytest.fixture
+def raster_save_path(tmp_path, boolean):
+    path = tmp_path / "raster.tif"
+    if boolean:
+        path = str(path)
+    return path
+
+
+def test_save(raster_save_path):
+    raster = testdata.raster.dem_small
+    raster.save(raster_save_path)
+    # Make sure round trip works
+    assert_rasters_equal(raster, Raster(raster_save_path))
 
 
 @pytest.mark.parametrize(
