### 2025-04-01 f3d80b3
Allow creation funcs to take dataarray objects as templates
