### 2025-04-01 5314003
Add copy kwarg to Raster.__array__
Closes #37
