### 2025-04-01 aea94a0
Revert "Bump version to v0.9.9"
This reverts commit 5fe25e934abcea66fa3dc1bb46c1bcbd22dc9f4b.
