+++ raster_tools/_version.py
@@ -1,1 +1,1 @@
-__version__ = "0.9.9"
+__version__ = "0.9.8"
