### 2025-04-01 4648b65
Add check for py 3.10+
