+++ raster_tools/_compat.py
@@ -1,4 +1,5 @@
 import os
+import sys
 
 import numpy as np
 
@@ -18,3 +19,5 @@ if gpd.options.use_pygeos:
 
 # Numpy 2.0 made several changes to type promotion rules.
 NUMPY_GE_2 = version_to_tuple(np.__version__) >= (2, 0, 0)
+
+PY_VER_310_PLUS = sys.version_info >= (3, 10)
