### 2025-04-29 #8134 @editorialbot
[REVIEW]: Raster Tools: An Open Source Toolbox for Raster Processing
**Submitting author:** @fbunt (<a href="http://orcid.org/0000-0003-4139-5355">Fredrick Bunt</a>)
**Repository:** https://github.com/UM-RMRS/raster_tools
**Branch with paper.md** (empty if default branch): main
**Version:** 0.9.11
