### 2025-06-22 #8134 @nagellette
@kanishkan91 I'm still reviewing this one. To provide some time to the author, there is a clear need for major improvement for the following under "Software" section in the checklist:

- **A statement of need:** Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work? - _There is a section providing a clarification about need but it doesn't contain anything about similar work. There are references to the libraries that are in use but it is not clear to me if there are any software/tool/library tackling similar problem._
- **State of the field:** Do the authors describe how this software compares to other commonly-used packages? - _Similar to the above one there are libraries that this software is using though what is the state of this one when compared to the other software in the domain that is in target._

Quality of writing and references are ok in this version but should be checked once the above improvements took place.
