### 2025-07-05 UM-RMRS/raster_tools#38 @nagellette
Benchmarking on processing time and storage space
Is there any benchmark about the processing time and storage space reduction compared to the similar tools as claimed in the README.md? "Our library focuses on significantly reducing processing time and storage space".

Issue is related to https://github.com/openjournals/joss-reviews/issues/8134
