### 2025-07-10 #8134 @nagellette
@kanishkan91 I have completed my review. Here are my notes:

- As summarized in previous message, a lot to be done on the improvement of the paper content. I don't think it contains enough info and academic content that can be accepted as is. The writing quality in the initial form is good but should be checked once the paper is approved.
- The tool is functional as described though the claimed performance and storage improvement in the repository is not backed with benchmarks. The performance improvement is the main claim of this study and it should be supported well. I have created an issue in the repository addressing this. Since there is no detail about the claimed improvement points and the compared tools/software/etc. it is not possible for me to investigate such improvements in all repository by myself.
