### 2025-07-14 #8134 @fbunt
@kanishkan91 Yep! We are working on putting together benchmarks. We'll add more to the paper, too. That might not be until after the benchmarking is done, though. Thanks for reviewing, @nagellette .
