### 2025-07-18 406f7e7
Update pre-commit config
