+++ .pre-commit-config.yaml
@@ -1,7 +1,7 @@
 exclude: 'tests/data/'
 repos:
   - repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v4.4.0
+    rev: v5.0.0
     hooks:
       - id: check-yaml
       - id: end-of-file-fixer
