### 2025-07-18 1d5b00e
Fix typo in docstring
