### 2025-07-18 2b7713a
Blacklist new numpy ufuncs
