### 2025-07-18 b7a6f3b
Make backward compatible with older numpy
