### 2025-07-19 6c2211c
Pin odc-geo to avoid issues in python 3.9
ref: https://github.com/opendatacube/odc-geo/issues/237
