+++ requirements/default.txt
@@ -8,7 +8,7 @@ geopandas
 netcdf4
 numba
 numpy>=1.22
-odc-geo
+odc-geo<=0.4.11
 pandas
 pyogrio
 pyproj
