### 2025-10-07 #8134 @fbunt
@kanishkan91 Thank you for the feedback. A quick update: we found an issue in our benchmarking methodology that was making it look like raster_tools scaled much better than it actually does. The scaling is less than we had hoped for. We will have a section on it and a comparison with some other tools, but we are putting some more emphasis on how raster_tools bridges the gaps in the Python raster/GIS stack.

To your points: Yep, we can add that it. Thanks.
