### 2025-11-30 c169ffc
Update paper with new edits
