+++ paper/paper.bib
@@ -1,125 +1,3 @@
-@article{qcrunch,
-    title = {The Quant Crunch: How the demand for data science skills is
-            disrupting the job market},
-    author = {Miller, Steven and Hughes, Debbie},
-    journal = {Business Higher Education Forum},
-    year = {2017}
-}
-
-@article{see2024,
-    author = {See, Linda and Lesiv, Myroslava and Schepaschenko, Dmitry},
-    title = {Integrating Remote Sensing and Geospatial Big Data for Land Cover
-            and Land Use Mapping and Monitoring},
-    journal = {Land},
-    volume = {13},
-    year = {2024},
-    number = {6},
-    article-number = {769},
-    url = {https://www.mdpi.com/2073-445X/13/6/769},
-    issn = {2073-445X},
-    abstract = {The last few decades have seen an explosion in the availability
-               of remotely sensed and geospatial big data, which are defined by the 3 Vs:
-               a large volume of data; a variety of different forms of data; and the rapid
-               velocity of data arrival [...]},
-    doi = {10.3390/land13060769}
-}
-
-@article{era5,
-    author = {Hersbach, Hans and Bell, Bill and Berrisford, Paul and Hirahara,
-             Shoji and Horányi, András and Muñoz-Sabater, Joaquín and Nicolas, Julien
-             and Peubey, Carole and Radu, Raluca and Schepers, Dinand and Simmons,
-             Adrian and Soci, Cornel and Abdalla, Saleh and Abellan, Xavier and Balsamo,
-             Gianpaolo and Bechtold, Peter and Biavati, Gionata and Bidlot, Jean and
-             Bonavita, Massimo and De Chiara, Giovanna and Dahlgren, Per and Dee, Dick
-             and Diamantakis, Michail and Dragani, Rossana and Flemming, Johannes and
-             Forbes, Richard and Fuentes, Manuel and Geer, Alan and Haimberger, Leo and
-             Healy, Sean and Hogan, Robin J. and Hólm, Elías and Janisková, Marta and
-             Keeley, Sarah and Laloyaux, Patrick and Lopez, Philippe and Lupu, Cristina
-             and Radnoti, Gabor and de Rosnay, Patricia and Rozum, Iryna and Vamborg,
-             Freja and Villaume, Sebastien and Thépaut, Jean-Noël},
-    title = {The ERA5 global reanalysis},
-    journal = {Quarterly Journal of the Royal Meteorological Society},
-    volume = {146},
-    number = {730},
-    pages = {1999-2049},
-    keywords = {climate reanalysis, Copernicus Climate Change Service, data
-               assimilation, ERA5, historical observations},
-    doi = {https://doi.org/10.1002/qj.3803},
-    url = {https://rmets.onlinelibrary.wiley.com/doi/abs/10.1002/qj.3803},
-    eprint = {https://rmets.onlinelibrary.wiley.com/doi/pdf/10.1002/qj.3803},
-    abstract = {Abstract Within the Copernicus Climate Change Service (C3S),
-               ECMWF is producing the ERA5 reanalysis which, once completed, will embody a
-               detailed record of the global atmosphere, land surface and ocean waves from
-               1950 onwards. This new reanalysis replaces the ERA-Interim reanalysis
-               (spanning 1979 onwards) which was started in 2006. ERA5 is based on the
-               Integrated Forecasting System (IFS) Cy41r2 which was operational in 2016.
-               ERA5 thus benefits from a decade of developments in model physics, core
-               dynamics and data assimilation. In addition to a significantly enhanced
-               horizontal resolution of 31 km, compared to 80 km for ERA-Interim, ERA5 has
-               hourly output throughout, and an uncertainty estimate from an ensemble
[…]
