+++ paper/paper.md
@@ -33,62 +33,261 @@ bibliography: paper.bib
 
 # Summary
 
-The Raster Tools Python package provides a homogeneous application programmers
-interface (API) for building scalable raster data processing pipelines.
-Pipelines built with Raster Tools can efficiently process extremely large
-raster datasets, including those larger than the system’s available memory.
-Raster Tools can be deployed on systems ranging from small laptops to high
-performance computing environments. To aid in the construction of raster
-processing pipelines, Raster Tools provides a consistent API to a suite of
-scalable raster processing functions including focal, zonal, clipping,
-convolution, and distance analysis operations.
+The Raster Tools Python package provides an application programmers interface
+(API) for building raster data processing pipelines. Pipelines built with
+Raster Tools can efficiently process extremely large raster datasets, including
+those larger than the system’s available memory. Raster Tools can be deployed
+on systems ranging from small laptops to high performance computing
+environments. To aid in the construction of raster processing pipelines, Raster
+Tools provides a consistent API to a suite of scalable raster processing
+functions including focal, zonal, clipping, convolution, and distance analysis
+operations.
 
 # Statement of Need
 
-In recent years, geospatial analysis has experienced a dramatic shift driven by
-the proliferation of global monitoring programs, new platforms for collecting
-geographically based observations, and advancements in remote sensing
-technology [@see2024]. This surge in data acquisition has led to the generation
-of massive raster datasets characterized by high spatial resolutions, dense
-temporal measurements, increased coverage of the electromagnetic spectrum, and
-greater spatial extents. Collectively, these mark a significant departure from
-the data volumes of the past [@qcrunch]. For example, MODIS satellites [@modis]
-are now acquiring 100 m resolution data and SENTINEL-2 [@sentinel2] is
-collecting imagery at 10-60 meter spatial resolution, globally at daily and
-five day intervals, respectively. The volume of such datasets has posed
-unprecedented challenges in terms of processing and analysis. In addition,
-there are numerous important products derived from primary data sources
-characterized by similar data volumes. Examples include interferometric
-synthetic radar estimates of velocities, soil moisture estimates, and
-vegetation indices. A specific example is the widely used ERA5 global
-reanalysis dataset, which now has a data catalog of roughly 5 petabytes
-[@era5].
+Python has become very popular as a language for data processing. This is
+because it provides a rich, general purpose, data stack, and simple, easy to
+understand syntax. Because of this, it is a good choice for tackling the
+expanding data problem. Some key libraries in the stack are Numpy [@numpy],
+Xarray [@xarray], SciPy [@scipy], and Dask [@dask]. Numpy forms the foundation
+of Python's stack by providing a fast and powerful N-dimensional (ND) array
+implementation and API for manipulating them. The Xarray package builds on this
+by supplying an API for working with labelled ND datasets. This is especially
+relevant to raster data with labeled dimensions and coordinates. Its data
+structures can be backed by any array object that implements Numpy's array
+interface, with Numpy being the default. The SciPy package provides optimized
+implementations of a wide variety of scientific and engineering algorithms.
 
-To meet modern data challenges, Python provides a rich, general purpose, data
-processing stack built on tools like Numpy [@numpy], SciPy [@scipy], Xarray
-[@xarray], and Dask [@dask]. With the use of GDAL [@gdal] bindings like
-Rasterio [@rasterio] and rioxarray [@rioxarray], this stack can be leveraged to
-carry out raster processing and geographic information systems (GIS) work. Yet,
-it is cumbersome and time consuming to develop pipelines that scale from small
-to extremely large raster datasets. This is for several primary reasons. First,
-many of the packages in Python’s data stack do not natively scale or only scale
-in a way that loads all data into memory, limiting the datasets sizes that can
[…]
