+++ paper/paper.bib
@@ -199,7 +199,7 @@
     editor = {Domingos Xavier Viegas and Luís Mário Ribeiro},
     title = "Advances in Fire Research 2022",
     booktitle = "Advances in Fire Research 2022",
-    publisher = {University of Coimbra Press}
+    publisher = {University of Coimbra Press},
     venue = {Coimbra, Portugal},
     month = oct,
     year = {2022},
