### 2025-11-30 e8bdf1e
Fix figure labeling and ref
