### 2025-11-30 #8134 @fbunt
@kanishkan91  I have committed the new version of the paper. It has been significantly altered. We included sections on scaling and research that has used our software. Hopefully these address your points above. I will work on README edits tomorrow and Monday.

The paper compiles but I'm having trouble getting the internal refs to work. I'll see if I can fix that tomorrow.
