### 2025-12-01 950d187
Remove redundant paper compile task
