+++ .github/workflows/paper-draft-pdf.yml
@@ -3,7 +3,7 @@ on:
   push:
     paths:
       - paper/**
-      - .github/workflows/draft-pdf.yml
+      - .github/workflows/paper-draft-pdf.yml
 
 jobs:
   paper:
