### 2025-12-02 3e6818f
Revamp readme intro, add install section, fix dependency list
