### 2026-01-05 #8134 @nagellette
@fbunt I'll reiterate my points in this note: https://github.com/openjournals/joss-reviews/issues/8134#issuecomment-2994403470

I understand about the clarification about the major low level libraries that raster-tools depend on such as numpy, scipy, xarray, etc. How these are related to your study and your reasioning is quite clear. As per my understanding, raster-tools remove low level complications these libraries introduce and provide a scalable, easy to use python interface to deal with raster data. On the other hand, it is still not clear if there are any other libraries that is targetting the "same need", resolving the same complications and how raster-tools add/exceeds to the state of their art. For example, I have came accross with xarray-spatial which sounds quite close to the raster-tools boilerplate functionality. whitebox-tools and geoutils can be checked as well.
