### 2026-01-10 e262535
Cleanup where testing
