### 2026-01-10 db1d0ba
Refactor to make assert_rasters_equal more robust and allow nv flexibility
