### 2026-01-10 5ca3d17
Fix null value corruption issue
Small float null values were being cast to float16 and loosing
precission. For instance, -9_999.0 was being cast and turning into
-10_000.0. This was due to the use of np.min_scalar_type being used when
checking the null value compatibility. np.result_type is now used.
