+++ raster_tools/masking.py
@@ -70,9 +70,10 @@ def reconcile_nullvalue_with_dtype(null_value, dtype, warn=False):
         return get_default_null_value(dtype)
 
     dtype = np.dtype(dtype)
-    null_value = np.min_scalar_type(null_value).type(null_value)
+    original_nv = null_value
+    null_value = np.result_type(null_value).type(null_value)
     if np.can_cast(null_value, dtype):
-        return null_value
+        return original_nv
 
     original_nv = null_value
     # TODO: stop checking if the value is an integer. Too much complexity. If
