### 2026-01-10 6a6c7e4
Allow a new null value to be specified in Raster.astype
