### 2026-01-15 #8134 @kanishkan91
@fbunt any updates on this? Based on @nagellette's queries I think it would be worth it to add a discussion comparing to similar packages or as the reviewer suggests-" other libraries that is targetting the "same need", resolving the same complications and how raster-tools add/exceeds to the state of their art."
