### 2026-01-22 b4d7f5c
Add new pub in Fire (Wiard 2026)
