+++ paper/paper.bib
@@ -187,6 +187,41 @@
     url = {https://scholarworks.umt.edu/etd/12531}
 }
 
+@article{wiard2026,
+    author = {Wiard-Greene, Lindsay and Johnson, Jesse and Hogland, John and Bunt, Fredrick and Bova, Jake},
+    title = {Investigating the Impact of Aerial Firefighting on Rate of Wildfire Spread},
+    journaL = {Fire},
+    volume = {9},
+    year = {2026},
+    number = {1},
+    article-number = {2},
+    url = {https://www.mdpi.com/2571-6255/9/1/2},
+    issn = {2571-6255},
+    abstract = {Aerial retardant drops are widely used in wildfire suppression,
+        yet their effectiveness in slowing fire spread remains difficult to
+        quantify at scale. This study evaluates their impact on wildfire
+        rate of spread (ROS) using a framework that combines observed and
+        counterfactual (synthetic) drop locations from 62 Oregon wildfires.
+        Synthetic drops were generated to simulate a no-suppression
+        baseline, enabling comparison of ROS with and without suppression.
+        We trained two random forest classifiers: one with real and
+        synthetic drops (full model) and one with only synthetic drops.
+        Both incorporated environmental and topographic features to predict
+        whether spread slowed following a drop. While the full model
+        performed well, the real-synthetic indicator had low feature
+        importance, offering limited causal evidence that aerial
+        suppression consistently reduced spread. The synthetic-only model
+        produced similar performance, suggesting that drops with observed
+        ROS reductions often coincided with favorable environmental and
+        topographic conditions and may have occurred independent of
+        suppression. These findings highlight the challenges of evaluating
+        suppression at scale and emphasize the need for finer data,
+        detailed operational records, and advanced modeling to better assess
+        the role of aerial fire retardant drops in future wildfire
+        management activities.},
+    doi = {10.3390/fire9010002}
+}
+
 @inproceedings{johnson2022,
     author = {Johnson, Jesse and Marcozzi, Anthony and Bunt, Frederick and Bova, Jacob and Hogland, John},
     title = {Predicting fire severity in Montana using a random forest classification scheme},
