### 2026-01-28 7bb53ea
Drop readthedocs config since we use github
