+++ .readthedocs.yml
@@ -1,14 +1,0 @@
-version: 2
-
-build:
-  os: ubuntu-22.04
-  tools:
-    python: mambaforge-4.10
-
-conda:
-  environment: requirements/docs.yml
-
-sphinx:
-  fail_on_warning: false
-
-formats: []
