### 2026-01-30 cfa3034
Incorporate changes based on reviewer comments
