+++ paper/paper.md
@@ -67,8 +67,8 @@ working with geospatial data, but does not implement many of GDAL's features. A
 third package, rioxarray [@rioxarray], is a plugin for Xarray that allows
 raster formats to be easily read into Xarray's data structures and for those
 data structures to use Rasterio operations on the underlying data. Finally,
-odc-geo [@odcgeo] provides several geospatial tools and algorithms such as
-grid definitions and reprojection. It also can function as a plugin for Xarray.
+odc-geo [@odcgeo] provides several geospatial tools and algorithms such as grid
+definitions and reprojection. It also can function as a plugin for Xarray.
 
 Together, these packages and the wider Python data stack can be leveraged to
 carry out simple and advanced raster processing and geographic information
@@ -79,14 +79,14 @@ while in much of Python’s data stack, this is handled with NaN values. These
 two paradigms can be difficult to reconcile, and it requires boilerplate code
 to properly mask data. Second, Numpy does not support parallel processing
 without writing C extensions, which is difficult at best. Newer packages such
-as Numba [@numba] and Numexpr [@numexpr] can help with this, but their scope
-is still limited. They also require more boilerplate code to be written. Third,
-if the size of the target dataset approaches or exceeds the available memory,
+as Numba [@numba] and Numexpr [@numexpr] can help with this, but their scope is
+still limited. They also require more boilerplate code to be written. Third, if
+the size of the target dataset approaches or exceeds the available memory,
 Numpy and everything built on it will exhaust the available memory and fail.
-This is because Numpy and most of the Python data stack is "in-core". All the
+This is because Numpy and most of the Python data stack is "in-core." All the
 data is in memory when processing occurs. To tackle this problem, more
-boilerplate code has to be written to manually chunk up the data and process it
-in batches.
+boilerplate code must be written to manually chunk the data and process it in
+batches.
 
 The last two issues can be solved by using the Dask package. Dask provides a
 relatively simple interface for parallel computing in Python. It allows large
@@ -105,6 +105,47 @@ compatibility gaps, such as making Rasterio and SciPy functions compatible with
 Dask. Code is also needed to properly handle missing data and to implement
 common and advanced GIS operations. There is no package that does all of this
 for the user.
+
+
+# Current State of the Field
+
+Several open-source Python packages try to address these needs in the Python
+ecosystem, but each does so differently and has its own limitations. Some
+examples are WhteboxTools [@wbtgat; @wbtman], GeoUtils [@geoutils], and
+Xarray-Spatial [@xrspatial].
+
+WhiteboxTools is a geospatial library providing hundreds of raster and GIS
+analysis functions through Python bindings. It is compiled and highly
+performant and capable of processing raster data in parallel. However, it
+presents several limitations: it lacks integration with the Python data stack
+and requires data to be fully loaded into memory for processing. Additionally,
+operations cannot be chained in memory; each tool must read from storage,
+process the data, and write the output back to storage before the next step can
+begin. A newer adaptation, Whitebox Workflows for Python, has been released,
+which allows chaining. It is proprietary freeware with a paid upgrade package,
+however.
+
+Built on top of Rasterio, GeoUtils is a package designed for efficient
+geospatial analysis with Python operator overloading and NumPy integration. It
+leverages Python’s multiprocessing module to process rasters in parallel.
+However, the package has notable limitations: most operations are eager and
+load the full dataset into memory, and it only implements a few geospatial
+operations such as reproject and proximity. Additionally, GeoUtils currently
+lacks compatibility with Dask and offers limited integration with Xarray. Work
[…]
