+++ paper/paper.bib
@@ -152,6 +152,40 @@
     doi = {10.5281/zenodo.2483274}
 }
 
+@article{wbtgat,
+    title = {Whitebox GAT: A case study in geomorphometric analysis.},
+    author = {J. B. Lindsay},
+    journal = {Computers & Geosciences},
+    year = {2016},
+    volume = {95},
+    pages = {75-84},
+    url = {http://dx.doi.org/10.1016/j.cageo.2016.07.003},
+}
+
+@manual{wbtman,
+    author = {Lindsay, John},
+    title = {WhiteboxTools User Manual}
+    year = {2018},
+    month = {03},
+    url = {https://www.whiteboxgeo.com/manual/wbt_book/preface.html},
+}
+
+@misc{geoutils,
+  author = {GeoUtils developers},
+  title = {GeoUtils: Consistent Geospatial Analysis in Python.},
+  year = {2020--2026},
+  url = {https://github.com/GlacioHack/geoutils},
+  note = {Accessed: 2026-01-16}
+}
+
+@misc{xrspatial,
+  author = {Xarray-Spatial developers},
+  title = {Xarray-Spatial},
+  year = {2020--2026},
+  url = {https://github.com/xarray-contrib/xarray-spatial},
+  note = {Accessed: 2026-01-15}
+}
+
 @mastersthesis{wiard2025,
     title = {Investigating the impact of aerial firefighting on rate of wildfire spread},
     author = {Wiard, Lindsay Ann},
