### 2026-01-31 1726dc2
Cite disributed and capitalize
