+++ paper/paper.bib
@@ -186,6 +186,14 @@
   note = {Accessed: 2026-01-15}
 }
 
+@manual{dask_dist,
+    title = {Dask Distributed Documentation},
+    author = {{Dask developers}},
+    year = {2026},
+    url = {https://distributed.dask.org/},
+    note = {Accessed: 2026-01-20}
+}
+
 @mastersthesis{wiard2025,
     title = {Investigating the impact of aerial firefighting on rate of wildfire spread},
     author = {Wiard, Lindsay Ann},
