### 2026-02-02 f93011e
Add cupy citation
