+++ paper/paper.bib
@@ -143,15 +143,6 @@
     year={2015}
 }
 
-@article{numexpr,
-    author = {McLeod, Robert},
-    year = {2018},
-    month = {12},
-    pages = {},
-    title = {NumExpr: Fast numerical expression evaluator for NumPy},
-    doi = {10.5281/zenodo.2483274}
-}
-
 @article{wbtgat,
     title = {Whitebox GAT: A case study in geomorphometric analysis.},
     author = {J. B. Lindsay},
@@ -192,6 +183,14 @@
     year = {2026},
     url = {https://distributed.dask.org/},
     note = {Accessed: 2026-01-20}
+}
+
+@inproceedings{cupy,
+  author       = "Okuta, Ryosuke and Unno, Yuya and Nishino, Daisuke and Hido, Shohei and Loomis, Crissman",
+  title        = "CuPy: A NumPy-Compatible Library for NVIDIA GPU Calculations",
+  booktitle    = "Proceedings of Workshop on Machine Learning Systems (LearningSys) in The Thirty-first Annual Conference on Neural Information Processing Systems (NIPS)",
+  year         = "2017",
+  url          = "http://learningsys.org/nips17/assets/papers/paper_16.pdf"
 }
 
 @mastersthesis{wiard2025,
