### 2026-02-04 7dc0470
Add rewrite of limitations section
