### 2026-02-05 00bb2d3
Add model predict paragraph
