### 2026-02-05 fbe7ceb
Update comparisons and add paragraph about Model Predict API
