+++ paper/paper.bib
@@ -155,7 +155,7 @@
 
 @manual{wbtman,
     author = {Lindsay, John},
-    title = {WhiteboxTools User Manual}
+    title = {WhiteboxTools User Manual},
     year = {2018},
     month = {03},
     url = {https://www.whiteboxgeo.com/manual/wbt_book/preface.html},
