### 2026-02-05 #8134 @fbunt
> I understand about the clarification about the major low level libraries that raster-tools depend on such as numpy, scipy, xarray, etc. How these are related to your study and your reasioning is quite clear. As per my understanding, raster-tools remove low level complications these libraries introduce and provide a scalable, easy to use python interface to deal with raster data. On the other hand, it is still not clear if there are any other libraries that is targetting the "same need", resolving the same complications and how raster-tools add/exceeds to the state of their art. For example, I have came accross with xarray-spatial which sounds quite close to the raster-tools boilerplate functionality. whitebox-tools and geoutils can be checked as well.

@nagellette  Thank you for the comments. We have added a section discussing other packages in the space trying to fill the same or similar needs. The packages you listed ended up being the major ones so those are what we discuss. In addition, we added a short section at the end discussing the comparison between them and raster-tools. We feel that raster-tools fills more of the gaps in the Python ecosystem while the others force trade-offs, thus adding to the state of the art. I hope we have made that clear in the updates.

Since we first submitted and did the benchmarking, the incompatibility between Dask's Distributed scheduler and one of our dependencies has been fixed fixed. We updated the scaling and limitations sections to reflect this but did not rerun the benchmarks with the new scheduler. Instead, we indicate that the results should be taken as a baseline with performance expected to improve with the new scheduler.

A few more edits were made throughout to increase clarity and cleanup language. We also added a newly released publication (Wiard 2026), which used raster-tools.

Thank you.
