+++ paper/paper.md
@@ -27,7 +27,7 @@ affiliations:
   - index: 2
     name: Rocky Mountain Research Station, USA
     ror: 04347cr60
-date: 10 January 2025
+date: 5 February 2026
 bibliography: paper.bib
 ---
 
