### 2026-02-19 6810db8
Assert rasters are valid
