### 2026-02-20 #8134 @fbunt
@kanishkan91 Thanks. Here you go.

* Version: 0.9.10
   - Tag: v0.9.10-joss
* DOI: https://doi.org/10.5281/zenodo.18676403
