### 2026-02-20 925e660
Fix read-only array error caused by changes in pandas 3
See: https://pandas.pydata.org/docs/user_guide/copy_on_write.html#read-only-numpy-arrays
pandas.Series.to_numpy() now returns read-only arrays.
