### 2026-02-20 3e9cf02
Update tests to run weekly
