### 2026-02-23 #8134 @kanishkan91
@fbunt  I am recommending this for acceptance now. I will be reading through the paper for other typos etc. shortly. The AEiC in this submission track will review shortly and if all goes well this will go live soon! Big thank you to @nagellette for reviewing! JOSS is volunteer run and relies heavily on researchers such as yourself.
