### 2026-02-23 9148597
Fix missing bib entry and typos in ref names
