+++ paper/paper.bib
@@ -143,6 +143,16 @@
     year={2015}
 }
 
+
+@article{numexpr,
+    author = {McLeod, Robert},
+    year = {2018},
+    month = {12},
+    pages = {},
+    title = {NumExpr: Fast numerical expression evaluator for NumPy},
+    doi = {10.5281/zenodo.2483274}
+}
+
 @article{wbtgat,
     title = {Whitebox GAT: A case study in geomorphometric analysis.},
     author = {J. B. Lindsay},
