### 2026-02-23 f5d1942
Fix missing 'Figure' before label ref
