### 2026-02-23 #8134 @fbunt
@kanishkan91 Sorry about that and thank you for catching it. I have fixed them now. How does this work with the zenodo archive? Do I need to make another? We have also released a new version (0.9.11) since the previous paper update. Thanks.
