### 2026-02-24 #8134 @fbunt
@kanishkan91 does the XML metadata error above indicate I need to fix something?
