### 2026-03-03 #8134 @fbunt
@kanishkan91 sounds good, thanks. Since We have made an update to the repo and paper, do we need to just update the zenodo archive or make a new one? Thanks.
