### 2026-03-04 #8134 @arfon
@fbunt – would you mind trying to edit the table in your paper so the last line of it is:

```
+--------------------------+------------------------------+--------------------------------------------------+
```
