### 2026-03-05 #8134 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
