### 2026-03-14 #8134 @crvernon
@fbunt - this submission is looking good.  I just need you to address the following before I accept for publication:

In your archive:

- [x] The latest version of your software on GitHub as declared in this thread is 0.9.11; however, your Zenodo archive points to 0.9.10.  Please ensure that the Zenodo archive release matches your specified release in this thread.

In your paper:

- [x] LINE 89:  "cuda" should be written in all caps as "CUDA"
- [x] LINE 182:  The in-text citation "(Connor et al., 2017), (O’Connor et al., 2022)" should appear as "Connor et al., 2017; O’Connor et al., 2022)".
- [x] LINE 223:  You are missing section header "References" in between your Conclusion section and the references.
- [x] LINE 235:  The "p" in "python" should be capitalized.  This can be done in your bib file by adding curly brackets around the characters you wish to maintain the formatting of.
- [x] LINE 256:  Capitalization should be like "Numba: A LLVM-based Python JIT compiler."

Let me know when these changes have been made and I'll accept.
