### 2026-03-17 2340f43
Fix copy issues
