+++ paper/paper.bib
@@ -136,7 +136,7 @@
 }
 
 @inproceedings{numba,
-    title={Numba: A llvm-based python jit compiler},
+    title={Numba: A LLVM-based Python JIT Compiler},
     author={Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
     booktitle={Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC},
     pages={1--6},
@@ -173,7 +173,7 @@
 
 @misc{geoutils,
   author = {GeoUtils developers},
-  title = {GeoUtils: Consistent Geospatial Analysis in Python.},
+  title = {GeoUtils: Consistent Geospatial Analysis in {P}ython.},
   year = {2020--2026},
   url = {https://github.com/GlacioHack/geoutils},
   note = {Accessed: 2026-01-16}
