### 2026-03-17 4d178d3
Fix more copy issues
