+++ paper/paper.bib
@@ -73,7 +73,7 @@
 }
 
 @manual{rasterio,
-    title = {Rasterio: geospatial raster I/O for {Python} programmers},
+    title = {Rasterio: geospatial raster {I/O} for {Python} programmers},
     author = {Sean Gillies and others},
     organization = {Mapbox},
     year = {2013--},
@@ -173,7 +173,7 @@
 
 @misc{geoutils,
   author = {GeoUtils developers},
-  title = {GeoUtils: Consistent Geospatial Analysis in {P}ython.},
+  title = {GeoUtils: Consistent Geospatial Analysis in {Python}.},
   year = {2020--2026},
   url = {https://github.com/GlacioHack/geoutils},
   note = {Accessed: 2026-01-16}
