+++ paper/paper.md
@@ -27,7 +27,7 @@ affiliations:
   - index: 2
     name: Rocky Mountain Research Station, USA
     ror: 04347cr60
-date: 23 February 2026
+date: 17 March 2026
 bibliography: paper.bib
 ---
 
