### 2026-03-19 #8134 @crvernon
@fbunt you are still missing the capital "p" in "python" here:

LINE 257: Capitalization should be like "Numba: A LLVM-based Python JIT compiler."
