+++ paper/paper.bib
@@ -136,7 +136,7 @@
 }
 
 @inproceedings{numba,
-    title={Numba: A LLVM-based Python JIT Compiler},
+    title={Numba: A LLVM-based {Python} JIT Compiler},
     author={Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
     booktitle={Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC},
     pages={1--6},
